Windows部分の中身の実装

+++ vdl/Src/vdl/Window/Windows/CWindow.cpp
@@ -4,7 +4,7 @@ void CWindow::Initialize()
 {
   constexpr DWORD kStyle = WS_OVERLAPPEDWINDOW ^ WS_MAXIMIZEBOX ^ WS_THICKFRAME | WS_VISIBLE;
 
-  RECT Rect = { 0, 0, WindowSize_.x, WindowSize_.y };
+  RECT Rect = { 0, 0, static_cast<long>(WindowSize_.x), static_cast<long>(WindowSize_.y) };
   ::AdjustWindowRect(&Rect, kStyle, false);
 
   hWnd_ = CreateWindowA(Constants::kClassName,
